+++ src/features/landing-page/first-section/images/curriculum.jpeg
